+++ examples/blog/src/pages/reference/script.js
@@ -1,0 +1,5 @@
+import { mountDocsSidebar } from "@/utils/docs-sidebar.js";
+
+export function mounted({ ctx }) {
+  mountDocsSidebar(ctx);
+}